src/requirements.txt
# Core Flask dependencies
Flask==3.0.3
Werkzeug==3.0.4
Jinja2==3.1.4
click==8.1.7
itsdangerous==2.2.0
MarkupSafe==2.1.5

# Flask extensions
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Cognito==1.21

# JWT and authentication
PyJWT==2.9.0
cryptography==43.0.0

# AWS and cloud dependencies
boto3==1.35.3
botocore==1.35.3
aws-lambda-wsgi==0.0.6

# Payment processing
stripe==7.11.0

# Authentication and security
bcrypt==4.2.0
cryptography==43.0.0
PyJWT==2.9.0
cognitojwt==1.4.1
python-jose==3.3.0

# HTTP and networking
requests==2.32.3
urllib3==2.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
packaging==24.1
six==1.16.0

# Cryptography dependencies
cffi==1.17.0
pycparser==2.22
ecdsa==0.19.0
rsa==4.9
pyasn1==0.6.0

# Development and deployment
gunicorn==23.0.0
blinker==1.8.2

# Testing dependencies
pytest>=8.3.3
pytest-cov>=4.0.0
pytest-mock>=3.12.0
pytest-flask>=1.3.0
pytest-env>=1.1.5
responses>=0.24.1
factory-boy>=3.3.0
moto>=4.2.0

# AWS transfer utilities
s3transfer==0.10.2
jmespath==1.0.1
